repackage application
