--- src/main/kotlin/com/wire/backups/exports/android/database/model/ConversationMembers.kt
@@ -1,10 +1,0 @@
-package com.wire.backups.exports.android.database.model
-
-import org.jetbrains.exposed.sql.Table
-
-object ConversationMembers : Table("ConversationMembers") {
-    val userId = text("user_id")
-    val conversationId = text("conv_id") references Conversations.id
-
-    override val primaryKey = PrimaryKey(userId, conversationId)
-}
