--- src/main/kotlin/com/wire/backups/exports/android/database/model/Conversations.kt
@@ -1,9 +1,0 @@
-package com.wire.backups.exports.android.database.model
-
-import org.jetbrains.exposed.sql.Table
-
-object Conversations : Table("Conversations") {
-    val id = text("_id")
-    val name = text("name").nullable()
-    override val primaryKey = PrimaryKey(id)
-}
