+++ src/main/kotlin/com/wire/backups/exports/android/database/model/Message.kt
@@ -1,4 +1,4 @@
-package com.wire.backups.exports.android.database.v2.model
+package com.wire.backups.exports.android.database.model
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
