+++ src/main/kotlin/com/wire/backups/exports/android/database/model/Model.kt
@@ -1,0 +1,5 @@
+package com.wire.backups.exports.android.database.model
+
+internal interface Model {
+    val id: String
+}
