--- src/main/kotlin/com/wire/backups/exports/android/database/v2/converters/Utils.kt
@@ -1,15 +1,0 @@
-package com.wire.backups.exports.android.database.v2.converters
-
-import java.time.Instant
-import java.time.ZoneOffset
-import java.time.format.DateTimeFormatter
-
-/**
- * String date format in yyyy-MM-dd'T'HH:mm:ss.SSS'Z' format.
- */
-internal typealias ExportDate = String
-
-internal fun Long.toExportDateFromAndroid(): ExportDate = Instant.ofEpochMilli(this).let { dateFormatter.format(it) }
-
-private val dateFormatter = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'")
-    .withZone(ZoneOffset.UTC)
