--- src/main/kotlin/com/wire/backups/exports/android/database/v2/model/Model.kt
@@ -1,5 +1,0 @@
-package com.wire.backups.exports.android.database.v2.model
-
-internal interface Model {
-    val id: String
-}
