+++ src/main/kotlin/com/wire/backups/exports/ios/database/MessagesConverter.kt
@@ -1,10 +1,10 @@
 package com.wire.backups.exports.ios.database
 
-import com.wire.backups.exports.android.database.converters.toExportDateFromIos
 import com.wire.backups.exports.ios.database.config.IosDatabase
 import com.wire.backups.exports.ios.database.model.GenericMessageData
 import com.wire.backups.exports.ios.database.model.Messages
 import com.wire.backups.exports.ios.model.IosMessageDto
+import com.wire.backups.exports.ios.toExportDateFromIos
 import com.wire.backups.exports.utils.mapCatching
 import com.wire.backups.exports.utils.rowExportFailed
 import com.wire.backups.exports.utils.transactionsLogger
